fix(ssi): cleanup mutation filter interface
This commit cleans up the common and default mutation filter so it's
more readable and concise.
