+++ pkg/clusteragent/admission/mutate/common/filter_test.go
@@ -76,7 +76,7 @@ func TestFailingInjectionConfig(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			nsFilter, _ := NewMutationFilter(tt.instrumentationEnabled, tt.enabledNamespaces, tt.disabledNamespaces)
+			nsFilter, _ := NewDefaulFilter(tt.instrumentationEnabled, tt.enabledNamespaces, tt.disabledNamespaces)
 			require.NotNil(t, nsFilter, "we should always get a filter")
 
 			checkedNamespaces := map[string]bool{}
